+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Add validation for Okapi and IdP URLs. Fixes UIORG-43.
 * Ignore yarn-error.log file. STRIPES-517.
 * Location-Institution CRUD. Fixes UIORG-54.
+* Add Service Point CRUD. Fixes UIORG-50.
 
 ## [2.2.0](https://github.com/folio-org/ui-organization/tree/v2.2.0) (2017-09-01)
 [Full Changelog](https://github.com/folio-org/ui-organization/compare/v2.1.0...v2.2.0)